Go back to readAsBinaryString
It's a deprecated function, but MDN docs say:

"Note that this method was once removed from the File API specification, but re-introduced for backward compatibility."

https://developer.mozilla.org/en-US/docs/Web/API/FileReader/readAsBinaryString

I do not believe it is going away any time soon. readAsArrayBuffer
is considered best now, but every solution I've seen there is
converting to Uint8Array and then passing through a string function.

Which seems like the long way around.