Generic WG handler and universal Size and Posn
Add a generic WG handler when no custom one is found - simply
queries what data we have in refData.

Universal Size and Posn checks if they were requested and not set
in the response to be returned. If not found, will query the DOM
to find the size and position.