Keyboard controls, focussing for List
This is a little messy: we introduce the idea of now having three
'pointers' in to a List. It's current 'start' position, it's last
selected position and retain drag start position as-is. This
appears to be enough to reproduce most Windows behaviours.

Focussing also now happens when you finish a drag.