Try to catch suspected null dereference. Comment out console.logs
The null dereference is when we walk the tree to find a particular
ID. If a leaf is null or undefined, then we know there's nothing
to walk! The problem here is that typeof null === 'object' - so it
expects to be able to walk the keys of it.

I've hit issues where my Windows VM has slowed to a crawl whilst
debugging and this has been made worse by excessive logging in the
frontend. I've commented all console.log calls, but left them
available to be reinstated as desired. I also hit an issue where I
spuriously triggered too much logging to the console - enough for
the APL side of EWC to give up on a response and fail a message.

So we'll default to quiet for now and reintroduce essential logging
only as we go forward.